fix: remove v6_8 drop old rule conditions and stream filters
