+++ src/Core/Migration/V6_8/Migration1773829004RegisterProductStreamIndexer.php
@@ -10,23 +10,18 @@ use Shopware\Core\Framework\Migration\MigrationStep;
  * @internal
  */
 #[Package('inventory')]
-class Migration1773829003RemoveLegacyLineItemProductStatesRuleCondition extends MigrationStep
+class Migration1773829004RegisterProductStreamIndexer extends MigrationStep
 {
     public function getCreationTimestamp(): int
     {
-        return 1773829003;
+        return 1773829004;
     }
 
     public function update(Connection $connection): void
     {
-        $deleted = $connection->executeStatement(
-            'DELETE FROM `rule_condition` WHERE `type` = :legacyType',
-            ['legacyType' => 'cartLineItemProductStates']
-        );
-
-        if ($deleted > 0) {
-            $this->registerIndexer($connection, 'rule.indexer');
-        }
+        // Similar to \Shopware\Core\Migration\V6_8\Migration1763125892RemoveProductStatesColumn::update
+        // Re-register the rule indexer to ensure product stream which might depend on the removed column are re-indexed properly
+        $this->registerIndexer($connection, 'product_stream.indexer');
     }
 
     public function updateDestructive(Connection $connection): void
